% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apra_hospital_coverage}
\alias{apra_hospital_coverage}
\title{Hospital treatment coverage data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 190 rows and 10 columns.
}
\source{
Australian Prudential Regulation Authority (APRA),
accessed on 03/12/2022.
}
\usage{
apra_hospital_coverage
}
\description{
The data shows the quarterly hospital treatment coverage by state or territory
as insured persons as percentage of population from 1970s to September 2022.
}
\keyword{datasets}
