% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{au_covid}
\alias{au_covid}
\title{Covid-19 data for Australia}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1045 rows and 3 columns.
}
\source{
Data extracted from the \code{tidycovid19} package using
\code{download_merged_data(cache = TRUE)}. Accessed on 03/12/22.
}
\usage{
au_covid
}
\description{
The data is the merged data from the Johns Hopkins University CSSE team.
}
\keyword{datasets}
