% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casey_emissions}
\alias{casey_emissions}
\title{Summary residential community emissions}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 29 rows and 24 columns.
}
\source{
City of Casey \url{https://data.casey.vic.gov.au/}
}
\usage{
casey_emissions
}
\description{
The dataset contains residential energy (electricity and gas) consumption
by suburb and by customer as well as associated GHG emissions.
}
\keyword{datasets}
