% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_bpsex}
\alias{census_bpsex}
\title{Birth place by sex from the Australian Census}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 288 rows and 5 columns.
}
\source{
Australian Census 2011, 2016 and 2021, Table 09 from the
Australian Bureau of Statistics \url{www.abs.gov.au}. Accessed on
03/12/2022.
}
\usage{
census_bpsex
}
\description{
ABS data used with permission from the Australian Bureau of Statistics.
The data is taken from the datapack from the ABS website for the 2011,
2016, and 2021 Australian Census.
}
\keyword{datasets}
