% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epa_airquality}
\alias{epa_airquality}
\title{EPA Air Watch Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15013066 rows and 11 columns.
}
\source{
Environmental Protection Agency, Accessed on 03/12/2022.
\url{https://discover.data.vic.gov.au/dataset/epa-air-watch-all-sites-air-quality-hourly-averages-yearly}
}
\usage{
epa_airquality
}
\description{
The data shows air quality information measured hourly by EPA’s air monitoring
stations around Victoria.
}
\keyword{datasets}
